Fix missing epoch income at start of day
Change the income calculation to use only the end_balance so income in the first epoch of each day is not missed. Requires one extra day before the start timestamp but only has to select the day and end_balance rather than day, start_balance, and end_balance. EndBalance is then written to totalIncomePerDay[date][1] and to totalIncomePerDay[date2][0]. Finally the extra half days at the start and end are skipped when returning the History data.